+++ previous-versions/accounting/src/codataccounting/models/operations/download_invoice_attachment.py
@@ -11,11 +11,13 @@ from typing import Optional
 @dataclasses.dataclass
 class DownloadInvoiceAttachmentRequest:
     attachment_id: str = dataclasses.field(metadata={'path_param': { 'field_name': 'attachmentId', 'style': 'simple', 'explode': False }})
-    r"""Unique identifier for an attachment"""
+    r"""Unique identifier for an attachment."""
     company_id: str = dataclasses.field(metadata={'path_param': { 'field_name': 'companyId', 'style': 'simple', 'explode': False }})
+    r"""Unique identifier for a company."""
     connection_id: str = dataclasses.field(metadata={'path_param': { 'field_name': 'connectionId', 'style': 'simple', 'explode': False }})
+    r"""Unique identifier for a connection."""
     invoice_id: str = dataclasses.field(metadata={'path_param': { 'field_name': 'invoiceId', 'style': 'simple', 'explode': False }})
-    r"""Unique identifier for an invoice"""
+    r"""Unique identifier for an invoice."""
     
 
 
@@ -24,11 +26,14 @@ class DownloadInvoiceAttachmentRequest:
 @dataclasses.dataclass
 class DownloadInvoiceAttachmentResponse:
     content_type: str = dataclasses.field()
+    r"""HTTP response content type for this operation"""
     status_code: int = dataclasses.field()
-    data: Optional[bytes] = dataclasses.field(default=None)
+    r"""HTTP response status code for this operation"""
+    data: Optional[requests_http.Response] = dataclasses.field(default=None)
     r"""Success"""
     error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)
     r"""Your API request was not properly authorized."""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
+    r"""Raw HTTP response; suitable for custom response parsing"""