+++ previous-versions/accounting/src/codataccounting/models/operations/get_bill_payments.py
@@ -12,7 +12,9 @@ from typing import Optional
 @dataclasses.dataclass
 class GetBillPaymentsRequest:
     bill_payment_id: str = dataclasses.field(metadata={'path_param': { 'field_name': 'billPaymentId', 'style': 'simple', 'explode': False }})
+    r"""Unique identifier for a bill payment."""
     company_id: str = dataclasses.field(metadata={'path_param': { 'field_name': 'companyId', 'style': 'simple', 'explode': False }})
+    r"""Unique identifier for a company."""
     
 
 
@@ -21,11 +23,14 @@ class GetBillPaymentsRequest:
 @dataclasses.dataclass
 class GetBillPaymentsResponse:
     content_type: str = dataclasses.field()
+    r"""HTTP response content type for this operation"""
     status_code: int = dataclasses.field()
+    r"""HTTP response status code for this operation"""
     bill_payment: Optional[shared_billpayment.BillPayment] = dataclasses.field(default=None)
     r"""Success"""
     error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)
     r"""Your API request was not properly authorized."""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
+    r"""Raw HTTP response; suitable for custom response parsing"""