+++ previous-versions/accounting/src/codataccounting/models/operations/get_customer.py
@@ -12,7 +12,9 @@ from typing import Optional
 @dataclasses.dataclass
 class GetCustomerRequest:
     company_id: str = dataclasses.field(metadata={'path_param': { 'field_name': 'companyId', 'style': 'simple', 'explode': False }})
+    r"""Unique identifier for a company."""
     customer_id: str = dataclasses.field(metadata={'path_param': { 'field_name': 'customerId', 'style': 'simple', 'explode': False }})
+    r"""Unique identifier for a customer."""
     
 
 
@@ -21,11 +23,14 @@ class GetCustomerRequest:
 @dataclasses.dataclass
 class GetCustomerResponse:
     content_type: str = dataclasses.field()
+    r"""HTTP response content type for this operation"""
     status_code: int = dataclasses.field()
+    r"""HTTP response status code for this operation"""
     customer: Optional[shared_customer.Customer] = dataclasses.field(default=None)
     r"""Success"""
     error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)
     r"""Your API request was not properly authorized."""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
+    r"""Raw HTTP response; suitable for custom response parsing"""