+++ previous-versions/accounting/src/codataccounting/models/operations/get_journal.py
@@ -12,7 +12,9 @@ from typing import Optional
 @dataclasses.dataclass
 class GetJournalRequest:
     company_id: str = dataclasses.field(metadata={'path_param': { 'field_name': 'companyId', 'style': 'simple', 'explode': False }})
+    r"""Unique identifier for a company."""
     journal_id: str = dataclasses.field(metadata={'path_param': { 'field_name': 'journalId', 'style': 'simple', 'explode': False }})
+    r"""Unique identifier for a journal."""
     
 
 
@@ -21,11 +23,14 @@ class GetJournalRequest:
 @dataclasses.dataclass
 class GetJournalResponse:
     content_type: str = dataclasses.field()
+    r"""HTTP response content type for this operation"""
     status_code: int = dataclasses.field()
+    r"""HTTP response status code for this operation"""
     error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)
     r"""Your API request was not properly authorized."""
     journal: Optional[shared_journal.Journal] = dataclasses.field(default=None)
     r"""Success"""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
+    r"""Raw HTTP response; suitable for custom response parsing"""