+++ previous-versions/accounting/src/codataccounting/models/operations/refresh_company_info.py
@@ -12,6 +12,7 @@ from typing import Optional
 @dataclasses.dataclass
 class RefreshCompanyInfoRequest:
     company_id: str = dataclasses.field(metadata={'path_param': { 'field_name': 'companyId', 'style': 'simple', 'explode': False }})
+    r"""Unique identifier for a company."""
     
 
 
@@ -20,11 +21,14 @@ class RefreshCompanyInfoRequest:
 @dataclasses.dataclass
 class RefreshCompanyInfoResponse:
     content_type: str = dataclasses.field()
+    r"""HTTP response content type for this operation"""
     status_code: int = dataclasses.field()
+    r"""HTTP response status code for this operation"""
     dataset: Optional[shared_dataset.Dataset] = dataclasses.field(default=None)
     r"""Success"""
     error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)
     r"""Your API request was not properly authorized."""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
+    r"""Raw HTTP response; suitable for custom response parsing"""