+++ previous-versions/accounting/src/codataccounting/models/operations/upload_billcreditnote_attachment.py
@@ -20,8 +20,11 @@ class UploadBillCreditNoteAttachmentRequestBody:
 @dataclasses.dataclass
 class UploadBillCreditNoteAttachmentRequest:
     bill_credit_note_id: str = dataclasses.field(metadata={'path_param': { 'field_name': 'billCreditNoteId', 'style': 'simple', 'explode': False }})
+    r"""Unique identifier for a bill credit note."""
     company_id: str = dataclasses.field(metadata={'path_param': { 'field_name': 'companyId', 'style': 'simple', 'explode': False }})
+    r"""Unique identifier for a company."""
     connection_id: str = dataclasses.field(metadata={'path_param': { 'field_name': 'connectionId', 'style': 'simple', 'explode': False }})
+    r"""Unique identifier for a connection."""
     request_body: Optional[UploadBillCreditNoteAttachmentRequestBody] = dataclasses.field(default=None, metadata={'multipart_form': { 'file': True }, 'request': { 'media_type': 'multipart/form-data' }})
     
 
@@ -31,9 +34,12 @@ class UploadBillCreditNoteAttachmentRequest:
 @dataclasses.dataclass
 class UploadBillCreditNoteAttachmentResponse:
     content_type: str = dataclasses.field()
+    r"""HTTP response content type for this operation"""
     status_code: int = dataclasses.field()
+    r"""HTTP response status code for this operation"""
     error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)
     r"""Your API request was not properly authorized."""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
+    r"""Raw HTTP response; suitable for custom response parsing"""